Added root user check. Which you can toggle off.
User by default are not allowed to run the script as a root user.

However that can explicitaly be disabled by using the --unsafe option.

However note that variable parsing happens before the root check.
Otherwise we couldn't depend on the --unsafe flag as that needs to be
parsed.

Thus future developers should be aware that any logic in the command
parser can be affected by the privileges of a root user